requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
chromadb
openai
tiktoken
unstructured
python-magic
pdf2image
langchain-openai